last seen activity and predicted wakeup times
