+++ server/internal/server/server.go
@@ -1,8 +1,8 @@
 package server
 
 import (
-	"encoding/binary"
-	"encoding/json"
+	// "encoding/binary"
+	// "encoding/json"
 	"errors"
 	"os"
 	"os/signal"
@@ -13,13 +13,13 @@ import (
 	"tinygo.org/x/bluetooth"
 )
 
-const ISO8601 = "2006-01-02T15:04:05.999Z"
 const DEFAULT_GATEWAY_HTTP_ENDPOINT = "https://openphm.org/gateway_data"
 
 var adapter = bluetooth.DefaultAdapter
 
 var DATA_SERVICE_UUID = bluetooth.MustParseUUID("2deacc71-7b29-4ff4-8fc2-59461c7a73f5")
 var DEBUG_DATA_CHRC_UUID = bluetooth.MustParseUUID("ad690aaa-cfd4-4b4a-96a6-1110cb6782f6")
+var CONTROL_DATA_CHRC_UUID = bluetooth.MustParseUUID("0b7f9057-38ef-4db5-8e25-64bc66fb1963")
 
 var FLUX_DATA_CHRC_UUID = bluetooth.MustParseUUID("68e92ad3-0fb5-4c93-8b99-0d21771576fd")
 var RTD_DATA_CHRC_UUID = bluetooth.MustParseUUID("e64d1230-86ba-46aa-a62d-736d6f58226c")
@@ -52,6 +52,7 @@ var configStartSampleChar bluetooth.Characteristic
 var Gateway *model.Gateway
 
 // List of known sensors, displayed by GUI
+// FIXME: move these to the model package
 var Sensors *[]model.Sensor
 
 // List of devices flagged for collection
@@ -81,8 +82,21 @@ func Init(ss *[]model.Sensor, g *model.Gateway) error {
 				UUID:  DEBUG_DATA_CHRC_UUID,
 				Flags: bluetooth.CharacteristicWritePermission | bluetooth.CharacteristicWriteWithoutResponsePermission,
 				WriteEvent: func(client bluetooth.Connection, address string, offset int, value []byte) {
-					//handleData("debug", client, address, offset, value)
+					// Dump any data received into a bin file
 					handleDebugData(address, value)
+				},
+			},
+			{
+				UUID:  CONTROL_DATA_CHRC_UUID,
+				Flags: bluetooth.CharacteristicWritePermission | bluetooth.CharacteristicWriteWithoutResponsePermission,
+				WriteEvent: func(client bluetooth.Connection, address string, offset int, value []byte) {
+					// Sensor wants to start/end upload. To start upload sensor
+					// sends a header with expected data size and type. To end
+					// an upload sensor sends a flag, then server validates
+					// data. Maybe for the future we could support upload
+					// resuming in case of a crash?
+
+					// TODO: lol
 				},
 			},
 			{
@@ -155,7 +169,8 @@ func Init(ss *[]model.Sensor, g *model.Gateway) error {
 				ReadEvent: func(client bluetooth.Connection, address string, offset int) []byte {
 					for i, dev := range flaggedForCollect {
 						if dev == address {
-							// Remove flag
+							// Remove flag FIXME: this is just invalid code.
+							// Replace with a map for the nicer API
 							flaggedForCollect[i] = flaggedForCollect[len(flaggedForCollect)-1]
 							flaggedForCollect = flaggedForCollect[:len(flaggedForCollect)-1]
 							return []byte{0x1}
@@ -291,190 +306,4 @@ func handleDebugData(address string, value []byte) {
 			bufferCache[address] = value
 		}
 	}
-}
-
-/*
- * Receive a data upload from the sensor
- * Each sensor data type would have a dedicated characteristic
- */
-func handleData(dataType string, _ bluetooth.Connection, address string, _mtu int, value []byte) {
-	if len(value) == 0 {
-		out.Logger.Println("Zero byte array received from " + address + " handling data for " + dataType)
-		return
-	}
-
-	// Find sensor that is sending data
-	macAddress, _ := model.StringToMac(address)
-	var sensor *model.Sensor = nil
-	for i, s := range *Sensors {
-		if s.Mac == macAddress {
-			sensor = &(*Sensors)[i]
-			break
-		}
-	}
-	// Ensure sensor is permitted to send data
-	// TODO only devices that pair with gateway are allowed to access this chrc anyways
-	if sensor == nil {
-		// BUG the MAC address received here is reversed somehow...
-		out.Logger.Println("Device " + address + " tried to send data, but it is not paired with this gateway")
-		return
-	}
-
-	// Append data to total data transmission
-	transmitData, ok := savePacket(value, macAddress, dataType)
-	if !ok {
-		// incomplete data, keep waiting for more
-		return
-	}
-
-	// Done collecting data, serialize to json and attempt immediate transfer after
-	// BUG this is always vibration datatype.
-	out.Logger.Println("Received " + dataType + " data transmission from " + model.MacToString(macAddress) + " (" + sensor.Name + ")")
-
-	// Pick apart data and place into json structures
-	var measurements []map[string]interface{}
-	if dataType == "vibration" {
-		measurements = handleVibrationData(transmitData)
-	} else if dataType == "temperature" {
-		measurements = handleTemperatureData(transmitData)
-	} else if dataType == "audio" {
-		measurements = handleAudioData(transmitData)
-	} else {
-		out.Logger.Println("unknown data type", dataType)
-		return
-	}
-
-	jsonData, err := json.Marshal(measurements)
-	if err != nil {
-		out.Logger.Println("Error:", err)
-		return
-	}
-
-	// Upload to gateway
-	resp, err := sendMeasurements(jsonData, Gateway)
-
-	// catch marshal errors
-	if err != nil {
-		out.Logger.Println("Error:", err)
-		// Save to disk instead
-		if err := saveUnsentMeasurements(jsonData, transmitData.timestamp); err != nil {
-			out.Logger.Println("Error:", err)
-		}
-		return
-	}
-
-	// TODO: we could use this point to verify gateway settings
-
-	// Catch gateway response
-	if resp.StatusCode != 200 {
-		out.Logger.Println("Error sending data to server")
-		// Print out response error
-		body := make([]byte, resp.ContentLength)
-		defer resp.Body.Close()
-		resp.Body.Read(body)
-		out.Logger.Println(string(body))
-		// Save data (again?)
-		if err := saveUnsentMeasurements(jsonData, transmitData.timestamp); err != nil {
-			out.Logger.Println(err)
-		}
-		return
-	}
-
-	// Send everything else in store
-	sendUnsentMeasurements()
-}
-
-// Accelerometer json output
-func handleVibrationData(transmitData Transmission) []map[string]interface{} {
-	rawData := transmitData.packets
-	numberOfMeasurements := len(rawData) / 6 // 3 axes, 2 bytes per axis => 6 bytes per measurement
-	out.Logger.Println("Vibration data consists of", len(rawData), "bytes =", numberOfMeasurements, "measurements.")
-	x, y, z := make([]int16, numberOfMeasurements), make([]int16, numberOfMeasurements), make([]int16, numberOfMeasurements)
-	for i := 0; i < len(rawData); i += 6 {
-		//out.Logger.Println(rawData[i:i+6])
-		x[i/6] = int16(binary.LittleEndian.Uint16(rawData[i : i+2]))
-		y[i/6] = int16(binary.LittleEndian.Uint16(rawData[i+2 : i+4]))
-		z[i/6] = int16(binary.LittleEndian.Uint16(rawData[i+4 : i+6]))
-	}
-
-	measurements := []map[string]interface{}{}
-	measurements = append(measurements,
-		map[string]interface{}{
-			"sensor_id":          model.MacToString(transmitData.macAddress),
-			"time":               transmitData.timestamp,
-			"measurement_type":   transmitData.dataType,
-			"sampling_frequency": transmitData.samplingFrequency,
-			"axis":               "x",
-			"raw_data":           x,
-		},
-		map[string]interface{}{
-			"sensor_id":          model.MacToString(transmitData.macAddress),
-			"time":               transmitData.timestamp,
-			"measurement_type":   transmitData.dataType,
-			"sampling_frequency": transmitData.samplingFrequency,
-			"axis":               "y",
-			"raw_data":           y,
-		},
-		map[string]interface{}{
-			"sensor_id":          model.MacToString(transmitData.macAddress),
-			"time":               transmitData.timestamp,
-			"measurement_type":   transmitData.dataType,
-			"sampling_frequency": transmitData.samplingFrequency,
-			"axis":               "z",
-			"raw_data":           z,
-		},
-	)
-
-	return measurements
-}
-
-func handleTemperatureData(transmitData Transmission) []map[string]interface{} {
-	measurements := []map[string]interface{}{}
-
-	if len(transmitData.packets) == 2 {
-		temperature, err := parseTemperatureData(binary.LittleEndian.Uint16(transmitData.packets))
-		if err == nil {
-			measurements = append(measurements,
-				map[string]interface{}{
-					"sensor_id":          model.MacToString(transmitData.macAddress),
-					"time":               transmitData.timestamp,
-					"measurement_type":   transmitData.dataType,
-					"sampling_frequency": transmitData.samplingFrequency,
-					"raw_data":           temperature,
-				},
-			)
-		} else {
-			out.Logger.Println("Error:", err)
-		}
-	} else {
-		out.Logger.Println("Invalid temperature data received")
-	}
-
-	return measurements
-}
-
-func handleAudioData(transmitData Transmission) []map[string]interface{} {
-	measurements := []map[string]interface{}{}
-	if len(transmitData.packets)%3 == 0 {
-		numberOfMeasurements := len(transmitData.packets) / 3
-		amplitude := make([]int, numberOfMeasurements)
-		for i := 0; i < numberOfMeasurements; i++ {
-			// actually a 24 bit integer
-			//amplitude[i] = int(transmitData.packets[i*3]) | int(transmitData.packets[i*3+1])<<8 | int(transmitData.packets[i*3+2])<<16
-			amplitude[i] = int(transmitData.packets[i*3+2])<<16 | int(transmitData.packets[i*3+1])<<8 | int(transmitData.packets[i*3])
-		}
-
-		measurements = append(measurements,
-			map[string]interface{}{
-				"sensor_id":          model.MacToString(transmitData.macAddress),
-				"time":               transmitData.timestamp,
-				"measurement_type":   transmitData.dataType,
-				"sampling_frequency": transmitData.samplingFrequency,
-				"raw_data":           amplitude,
-			},
-		)
-	} else {
-		out.Logger.Println("Invalid audio data received. Packets of length", len(transmitData.packets), "not multiple of 3.")
-	}
-	return measurements
 }
